Fix Clang-18 double-promotion warnings in tests
Clang-18 is stricter than GCC about -Wdouble-promotion warnings.
Fixed implicit float-to-double conversions when comparing GLM vector
components (float) with doctest::Approx (double).

Changes:
- Added explicit static_cast<double>() for all glm::vec3/vec4 component
  comparisons with doctest::Approx
- Affects 16 CHECK statements across vector and matrix operation tests

This fixes CI failures on ubuntu-24.04 with clang-18 while maintaining
compatibility with GCC-13.

+++ tests/test_main.cpp
@@ -17,9 +17,9 @@ TEST_CASE("GLM vector operations")
         glm::vec3 b(4.0F, 5.0F, 6.0F);
         glm::vec3 result = a + b;
 
-        CHECK(result.x == doctest::Approx(5.0));
-        CHECK(result.y == doctest::Approx(7.0));
-        CHECK(result.z == doctest::Approx(9.0));
+        CHECK(static_cast<double>(result.x) == doctest::Approx(5.0));
+        CHECK(static_cast<double>(result.y) == doctest::Approx(7.0));
+        CHECK(static_cast<double>(result.z) == doctest::Approx(9.0));
     }
 
     SUBCASE("Vector dot product")
@@ -37,9 +37,9 @@ TEST_CASE("GLM vector operations")
         glm::vec3 b(0.0F, 1.0F, 0.0F);
         glm::vec3 result = glm::cross(a, b);
 
-        CHECK(result.x == doctest::Approx(0.0));
-        CHECK(result.y == doctest::Approx(0.0));
-        CHECK(result.z == doctest::Approx(1.0));
+        CHECK(static_cast<double>(result.x) == doctest::Approx(0.0));
+        CHECK(static_cast<double>(result.y) == doctest::Approx(0.0));
+        CHECK(static_cast<double>(result.z) == doctest::Approx(1.0));
     }
 }
 
@@ -51,10 +51,10 @@ TEST_CASE("GLM matrix operations")
         glm::vec4 point(1.0F, 2.0F, 3.0F, 1.0F);
         glm::vec4 result = identity * point;
 
-        CHECK(result.x == doctest::Approx(1.0));
-        CHECK(result.y == doctest::Approx(2.0));
-        CHECK(result.z == doctest::Approx(3.0));
-        CHECK(result.w == doctest::Approx(1.0));
+        CHECK(static_cast<double>(result.x) == doctest::Approx(1.0));
+        CHECK(static_cast<double>(result.y) == doctest::Approx(2.0));
+        CHECK(static_cast<double>(result.z) == doctest::Approx(3.0));
+        CHECK(static_cast<double>(result.w) == doctest::Approx(1.0));
     }
 
     SUBCASE("Translation matrix")
@@ -65,10 +65,10 @@ TEST_CASE("GLM matrix operations")
         glm::vec4 point(0.0F, 0.0F, 0.0F, 1.0F);
         glm::vec4 result = transform * point;
 
-        CHECK(result.x == doctest::Approx(5.0));
-        CHECK(result.y == doctest::Approx(10.0));
-        CHECK(result.z == doctest::Approx(15.0));
-        CHECK(result.w == doctest::Approx(1.0));
+        CHECK(static_cast<double>(result.x) == doctest::Approx(5.0));
+        CHECK(static_cast<double>(result.y) == doctest::Approx(10.0));
+        CHECK(static_cast<double>(result.z) == doctest::Approx(15.0));
+        CHECK(static_cast<double>(result.w) == doctest::Approx(1.0));
     }
 
     SUBCASE("Rotation matrix (90 degrees around Z-axis)")
@@ -79,10 +79,10 @@ TEST_CASE("GLM matrix operations")
         glm::vec4 point(1.0F, 0.0F, 0.0F, 1.0F);
         glm::vec4 result = transform * point;
 
-        CHECK(result.x == doctest::Approx(0.0).epsilon(0.0001));
-        CHECK(result.y == doctest::Approx(1.0).epsilon(0.0001));
-        CHECK(result.z == doctest::Approx(0.0).epsilon(0.0001));
-        CHECK(result.w == doctest::Approx(1.0).epsilon(0.0001));
+        CHECK(static_cast<double>(result.x) == doctest::Approx(0.0).epsilon(0.0001));
+        CHECK(static_cast<double>(result.y) == doctest::Approx(1.0).epsilon(0.0001));
+        CHECK(static_cast<double>(result.z) == doctest::Approx(0.0).epsilon(0.0001));
+        CHECK(static_cast<double>(result.w) == doctest::Approx(1.0).epsilon(0.0001));
     }
 }
